backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
httpx==0.28.1
python-dotenv==1.0.1
Pillow==10.4.0
google-genai==1.5.0
json-repair==0.30.3
PyMuPDF==1.24.10
